Further Organization of Character Scripts
Created a Base Character script to serve as the parent for Battle and
NPC character types. Generalized Dialog to NPC and a Base type, looking
to ensure that all types of characters are accounted for and that no
character has attributes or properties that they don't need.
