+++ Assets/Scenes/CharacterAttributeTesting.unity
@@ -217,6 +217,65 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 50, y: -30, z: 0}
+--- !u!1 &910080021
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 910080024}
+  - component: {fileID: 910080023}
+  - component: {fileID: 910080022}
+  m_Layer: 0
+  m_Name: NPCPlaceholder
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &910080022
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 910080021}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 837c9816b557fa34aa65dcf4ef42a5bd, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &910080023
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 910080021}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d777c84ff269c48459b67a639e60a670, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  dialogBehaviour: {fileID: 0}
+  dialogNodeGraph: {fileID: 0}
+--- !u!4 &910080024
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 910080021}
+  serializedVersion: 2
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1335428621
 GameObject:
   m_ObjectHideFlags: 0
@@ -227,7 +286,7 @@ GameObject:
   m_Component:
   - component: {fileID: 1335428624}
   - component: {fileID: 1335428623}
-  - component: {fileID: 1335428622}
+  - component: {fileID: 1335428625}
   m_Layer: 0
   m_Name: CharacterPlaceholder
   m_TagString: Untagged
@@ -235,21 +294,6 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!114 &1335428622
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1335428621}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3630121d8ea61774bae980aecf8950eb, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  attackAttribute: {fileID: 11400000, guid: 5d1dc61e5a8004c49a97dee563555f8b, type: 2}
-  defenseAttribute: {fileID: 11400000, guid: 094df27df26bd0140a3e7efa5faae3ff, type: 2}
-  speedAttribute: {fileID: 11400000, guid: b2c5de044aed80344ac1a6546ac7ad08, type: 2}
 --- !u!114 &1335428623
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -278,6 +322,21 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1335428625
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1335428621}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 39615a6890207874ab3f897701f4d2e2, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  attackAttribute: {fileID: 0}
+  defenseAttribute: {fileID: 0}
+  speedAttribute: {fileID: 0}
 --- !u!1 &1340454978
 GameObject:
   m_ObjectHideFlags: 0
@@ -377,3 +436,4 @@ SceneRoots:
   - {fileID: 1340454981}
   - {fileID: 473715934}
   - {fileID: 1335428624}
+  - {fileID: 910080024}
