+++ Assets/Scripts/Character/Attributes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 54d1e45fe78a15945a257a11f75a060e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
