+++ Assets/Scripts/Character/Attributes/Attribute.cs
