+++ Assets/Scripts/Character/Attributes/Dummy Data/Main Character Speed Data.asset
@@ -9,9 +9,9 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0ab971f51d9c69547aa80c8fdd59f112, type: 3}
-  m_Name: Main Character Defense Data
+  m_Script: {fileID: 11500000, guid: 32f78c23cef948740824b0817d12406b, type: 3}
+  m_Name: Main Character Speed Data
   m_EditorClassIdentifier: 
   values:
-  - level: 0
-    value: 0
+  - level: 1
+    value: 1
