+++ Assets/Scripts/Character/BattleCharacter.cs
@@ -1,0 +1,22 @@
+using UnityEngine;
+
+namespace SOS.AndrewsAdventure.Character
+{
+    [RequireComponent(typeof(Health))]
+    public class BattleCharacter : Character
+    {
+        [SerializeField] AttackAttribute attackAttribute;
+        [SerializeField] DefenseAttribute defenseAttribute;
+        [SerializeField] SpeedAttribute speedAttribute;
+
+        void Start()
+        {
+            print(GetData());
+        }
+
+        private string GetData()
+        {
+            return attackAttribute.GetData(1) + ", " + defenseAttribute.GetData(1) + ", " + speedAttribute.GetData(1);
+        }
+    }
+}
