+++ Assets/Scripts/Character/BattleCharacter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 39615a6890207874ab3f897701f4d2e2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
