+++ Assets/Scripts/Character/NPC.cs
@@ -1,0 +1,11 @@
+using SOS.AndrewsAdventure.Dialog;
+using UnityEngine;
+
+namespace SOS.AndrewsAdventure.Character
+{
+    [RequireComponent(typeof(NPCDialog))]
+    public class NPC : Character
+    {
+
+    }
+}
