+++ Assets/Scripts/Character/NPC.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 837c9816b557fa34aa65dcf4ef42a5bd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
