+++ Assets/Scripts/Dialogue/BaseDialog.cs
@@ -1,0 +1,16 @@
+using cherrydev;
+using UnityEngine;
+
+namespace SOS.AndrewsAdventure.Dialog
+{
+    public class BaseDialog : MonoBehaviour
+    {
+        [SerializeField] private DialogBehaviour dialogBehaviour;
+        [SerializeField] private DialogNodeGraph dialogNodeGraph;
+
+        void Start()
+        {
+            dialogBehaviour.StartDialog(dialogNodeGraph);
+        }
+    }
+}
