+++ Assets/Scripts/Dialogue/BaseDialog.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fd7639b45f5f9484ebe9c473a43a4f1f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
