+++ Assets/Scripts/Dialogue/NPCDialog.cs
@@ -1,0 +1,7 @@
+namespace SOS.AndrewsAdventure.Dialog
+{
+    public class NPCDialog : BaseDialog
+    {
+
+    }
+}
