+++ Assets/Scripts/Dialogue/NPCDialog.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d777c84ff269c48459b67a639e60a670
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
